refactor(shared): extract duplicated DTOs into Shared.Contracts
PagedResult<T>, StatisticsDTO/SubjectStat/TopicStat and UserInfo were
defined identically in both backend/api/DTOs and blazor/blazor/models.
Move the canonical definitions into the existing Shared class library
under Shared.Contracts and delete both copies. Consumers gain a
single 'using Shared.Contracts;' / '@using Shared.Contracts'.

The Shared project is already referenced by both api.csproj and
blazor.csproj, so no project-reference changes are needed.
