--- backend/api/DTOs/SubjectStat.cs
@@ -1,9 +1,0 @@
-namespace api.DTOs;
-
-public class SubjectStat
-{
-    public string Subject { get; set; } = "";
-    public int Count { get; set; }
-    public int TotalPoints { get; set; }
-    public double AveragePoints { get; set; }
-}
