--- backend/api/DTOs/TopicStat.cs
@@ -1,9 +1,0 @@
-namespace api.DTOs;
-
-public class TopicStat
-{
-    public string Topic { get; set; } = "";
-    public int Count { get; set; }
-    public int TotalPoints { get; set; }
-    public double AveragePoints { get; set; }
-}
