+++ backend/api/Services/AssignmentService.cs
@@ -2,6 +2,7 @@ using api.Data;
 using api.DTOs;
 using api.Models;
 using Microsoft.EntityFrameworkCore;
+using Shared.Contracts;
 
 namespace api.Services;
