+++ blazor/blazor/Pages/AssignmentAssignmentSheetOverview.razor
@@ -1,4 +1,5 @@
 @page "/assignment-assignment-sheet-overview"
+@using Shared.Contracts
 @using Shared
 @using blazor.models
 @using System.Net.Http.Json
