+++ blazor/blazor/Pages/AssignmentsStudent.razor
@@ -1,4 +1,5 @@
 @page "/assignments-student"
+@using Shared.Contracts
 @using Shared
 @using blazor.models
 @using System.Net.Http.Json
