+++ blazor/blazor/Pages/CreateAssignmentSheet.razor
@@ -1,4 +1,5 @@
 @page "/create-assignment-sheet"
+@using Shared.Contracts
 @using blazor.models
 @using Shared
 @using System.Net.Http.Json
