+++ blazor/blazor/Pages/EditAssignmentSheet.razor
@@ -1,4 +1,5 @@
 @page "/edit-assignment-sheet"
+@using Shared.Contracts
 @using Shared
 @using blazor.models
 @using System.Net.Http.Json
