+++ blazor/blazor/Pages/FilterAssignments.razor
@@ -1,4 +1,5 @@
 @page "/filter-assignments"
+@using Shared.Contracts
 @using Shared
 @using blazor.models
 @using System.Net.Http.Json
