+++ blazor/blazor/Pages/Statistics.razor
@@ -1,4 +1,5 @@
 @page "/statistik-elev"
+@using Shared.Contracts
 @using Shared
 @using blazor.models
 @using System.Globalization
