+++ blazor/blazor/Pages/StudentDashboard.razor
@@ -1,4 +1,5 @@
 @page "/student-dashboard"
+@using Shared.Contracts
 @using Shared
 @using blazor.models
 @using Microsoft.AspNetCore.Components.Web
