--- blazor/blazor/models/PagedResult.cs
@@ -1,10 +1,0 @@
-namespace blazor.models;
-
-public class PagedResult<T>
-{
-    public List<T> Items { get; set; } = new();
-    public int Page { get; set; }
-    public int PageSize { get; set; }
-    public int Total { get; set; }
-    public int TotalPages { get; set; }
-}
